+++ Commons/TransactionApi/src/main/java/com/mfino/transactionapi/handlers/nfc/impl/NFCPocketTopupInquiryHandlerImpl.java
@@ -259,7 +259,7 @@ public class NFCPocketTopupInquiryHandlerImpl  extends FIXMessageHandler impleme
 		// Saves the Transaction Id returned from Back End		
 		TransactionResponse transactionResponse = checkBackEndResponse(response);
 		if (transactionResponse.getTransactionId() !=null){
-			sctl.setTransactionid(BigDecimal.valueOf(transactionResponse.getTransactionId()));
+			sctl.setTransactionid(transactionResponse.getTransactionId());
 			bankAccountToBankAccount.setTransactionID(transactionResponse.getTransactionId());
 			result.setTransactionID(transactionResponse.getTransactionId());
 			transactionChargingService.saveServiceTransactionLog(sctl);
